Comparing two list before loading. ItemDetails rating update view. Sort items by rating.

+++ app/src/main/java/com/online/booking/data/viewmodel/ItemViewModel.kt
@@ -27,13 +27,11 @@ class ItemViewModel(application: Application) : AndroidViewModel(application) {
         try {
             val received = repository.getItems()
             //Compare two lists
-            //if( (received.size != items.size) || (received.size == items.size
-            //    && received.zip(items).all { (x, y) -> x == y })
-            //){
+            if( !isEqual(received.sortedBy { it.id }, items.sortedBy { it.id }) ){
                 itemDao.deleteAll()
                 received.forEach { item ->  itemDao.insert( item ) }
                 emit(Resource.remoteSuccess(data = received))
-            //}
+            }
         } catch (exception: Exception) {
             emit(Resource.error(data = items, message = exception.message ?: DEFAULT_ERR_MESSAGE))
         }
@@ -47,10 +45,10 @@ class ItemViewModel(application: Application) : AndroidViewModel(application) {
         }
         try {
             val received = repository.getItem( id )
-            //if(received != item){
+            if(received != item){
                 itemDao.update( received )
                 emit(Resource.remoteSuccess(data = received))
-            //}
+            }
         } catch (exception: Exception) {
             emit(Resource.error(data = item, message = exception.message ?: DEFAULT_ERR_MESSAGE))
         }
@@ -77,6 +75,15 @@ class ItemViewModel(application: Application) : AndroidViewModel(application) {
         } catch (exception: Exception){
             emit(Resource.error(data = null, message = exception.message ?: DEFAULT_ERR_MESSAGE))
         }
+    }
+
+    fun<T> isEqual(first: List<T>, second: List<T>): Boolean {
+
+        if (first.size != second.size) {
+            return false
+        }
+
+        return first.zip(second).all { (x, y) -> x == y }
     }
 
 }
